% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{ntwd_browse}
\alias{ntwd_browse}
\title{Quickly browse to Nationwide's webpage}
\usage{
ntwd_browse()
}
\description{
This function take you to the Nationwide's House Price Index webpage and
return the URL invisibly.
}
\examples{
ntwd_browse()
}
