% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{ntwd_dataset}
\alias{ntwd_dataset}
\title{Fetch house price data fro Nationwide}
\usage{
ntwd_dataset()
}
\description{
\itemize{
 \item UK Series
   \describe{
     \item{\strong{monthly}}{UK Monthly Indices (Post '91) (This link will open in a new window)}
     \item{\emph{quarterly}}{First item}
   }
}
}
\details{
\describe{
  \item{monthly}{UK Monthly Indices (Post '91)}
  \item{quarterly}{First item}
  \item{since-1952}{First item}
  \item{inflation-adjusted}{First item}
  \item{all-prop}{First item}
  \item{seasonal-regional}{First item}
  \item{new-prop}{First item}
  \item{mod-prop}{First item}
  \item{not-new-prop}{First item}
  \item{first-by}{First item}
  \item{fowner}{First item}
  \item{terraced}{First item}
  \item{flats}{First item}
  \item{semi-detached}{First item}
  \item{detached}{First item}
  \item{ftb-ind}{Affordability Estimates: First Time Buyer Affordability Indices}
  \item{ftb-hper}{Affordability Estimates: First Time Buyer House Price Earnings Ratios}
}
}
