% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ntwd_meta}
\alias{ntwd_meta}
\title{Access object's metadata}
\usage{
ntwd_meta(x)
}
\arguments{
\item{x}{object return from \code{\link{ntwd_get}}}
}
\description{
Some datasets in nationwide contain metadata that cannot be displayed in
a dataframe, hence using \code{ntwd_meta} the grants access to the
object's metadata that are stored as attributes to the dataframe.
}
\details{
Not all objects contain metadata
}
\examples{
x <- ntwd_get("since_1952")
ntwd_meta(x)
}
